Use Target Platform Definition DSL to generate target platforms
The "Target Platform Definition DSL and Generator" [1] heavily
simplifies maintenance of target platforms. It allows to modularize
target platform definitions which eliminates code duplication. The
.target files understood by P2 and Tycho are generated from .tpd files
which are written in the target platform definition DSL.

In order to edit .tpd files and generate .target files install the
"Target Platform Definition DSL and Generator" 2.0 or later [2]  (Note:
on Kepler you also need to add [3] to get Xtext 2.5 which is not
available by default on Kepler). This tools is needed only if you need
to change the Target Platform definition files (*.targetplatform and
*.tpd) and re-generate the *.target files. In normal development you do
not need this and can simply use the generated *.target themselves.

In addition
- update Orbit repository for 4.5 to Mars M5
- use latest released Orbit p2 repository for platform version Luna
  and earlier

[1] https://github.com/mbarbero/fr.obeo.releng.targetplatform
[2] http://mbarbero.github.io/fr.obeo.releng.targetplatform/p2/latest/
[3] http://download.eclipse.org/modeling/tmf/xtext/updates/releases/

Change-Id: Ia701972785c3e88aba66a7f15a2b3cf638727eea
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>