Revert "Config: Distinguish between empty and null strings"
This reverts commit 96eb3ee3976e7e9e3e118851fa614cce8a1f7d88, which
broke Gerrit tests that set a config value to 'null', serialize the
result, deserialize, and expect 'null' from Config.getString[1].

The intent of that commit was to make it possible to distinguish between
an absent and an empty config value, which we'll have to do with a new
method.

Revert the behavior change.  Keep the tests from 428cb23f2de8, since
they test the behavior more precisely than the old tests did.

[1] https://gerrit-review.googlesource.com/68452

Change-Id: Ie8042f380ea0e34e3203e1991aa0feb2e6e44641
Signed-off-by: Jonathan Nieder <jrn@google.com>